requirements.txt
Flask
python-dotenv
requests
sib-api-v3-sdk